# ApiDoc.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Api Call Center",
        "description": "Daftar Api Call Center",
        "version": "1.0.0"
    },
...
    "servers": [
        {
            "url": "http://192.168.3.98:8008",
            "description": "Server Staging"
        },
        {
            "url": "http://localhost:8008",
            "description": "Server Local"
        }
    ],
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT"
            }
        }
    },
...
    "paths": {
...
        "/emails": {
            "get": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": [
                    "Emails"
                ],
                "summary": "Get All Emails",
                "description": "",
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not found."
                    }
                }
            },
            "post": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": [
                    "Emails"
                ],
                "summary": "Insert Emails",
                "description": "",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "type": "object",
                                "required": [
                                    "direction_id",
                                    "mail_date",
                                    "email_from",
                                    "email_to",
                                    "subject",
                                    "content",
                                    "content_html"
                                ],
                                "properties": {
                                    "gtw_id": {
                                        "type": "string"
                                    },
                                    "is_blocking": {
                                        "type": "integer"
                                    },
                                    "direction_id": {
                                        "type": "integer"
                                    },
                                    "mail_date": {
                                        "type": "date"
                                    },
                                    "subject": {
                                        "type": "string"
                                    },
                                    "email_from": {
                                        "type": "string"
                                    },
                                    "email_to": {
                                        "type": "string"
                                    },
                                    "email_cc": {
                                        "type": "string"
                                    },
                                    "email_bcc": {
                                        "type": "string"
                                    },
                                    "content": {
                                        "type": "string"
                                    },
                                    "content_html": {
                                        "type": "string"
                                    },
                                    "note": {
                                        "type": "string"
                                    },
                                    "mail_error_info": {
                                        "type": "string"
                                    },
                                    "outbound_category_id": {
                                        "type": "integer"
                                    },
                                    "outbound_category_detail_id": {
                                        "type": "integer"
                                    },
                                    "is_complete": {
                                        "type": "integer"
                                    },
                                    "uid": {
                                        "type": "string"
                                    },
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "campaign_id": {
                                        "type": "integer"
                                    },
                                    "customer_id": {
                                        "type": "integer"
                                    },
                                    "read_at": {
                                        "type": "date"
                                    },
                                    "read_by": {
                                        "type": "integer"
                                    },
                                    "reply_id": {
                                        "type": "integer"
                                    },
                                    "reply_at": {
                                        "type": "date"
                                    },
                                    "reply_by": {
                                        "type": "integer"
                                    },
                                    "forward_at": {
                                        "type": "date"
                                    },
                                    "forward_by": {
                                        "type": "integer"
                                    },
                                    "created_at": {
                                        "type": "date"
                                    },
                                    "created_by": {
                                        "type": "integer"
                                    },
                                    "updated_at": {
                                        "type": "date"
                                    },
                                    "updated_by": {
                                        "type": "integer"
                                    },
                                    "cust_name": {
                                        "type": "string"
                                    },
                                    "inbound_type_detail": {
                                        "type": "string"
                                    },
                                    "source_media_id": {
                                        "type": "integer"
                                    },
                                    "send_error_attempt": {
                                        "type": "string"
                                    },
                                    "is_spam": {
                                        "type": "integer"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not found."
                    }
                }
            }
        },
        "/emails/{id}": {
            "get": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": [
                    "Emails"
                ],
                "summary": "Get By Id Emails",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "type": "integer",
                        "required": true,
                        "description": "Numeric ID of the Emails to get."
                    }
                ],
                "description": "",
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not found."
                    }
                }
            },
            "put": {
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "tags": [
                    "Emails"
                ],
                "summary": "Update By Id Emails",
                "parameters": [
                    {
                        "in": "path",
                        "name": "id",
                        "type": "integer",
                        "required": true,
                        "description": "Numeric ID of the Emails to update."
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/x-www-form-urlencoded": {
                            "schema": {
                                "type": "object",
                                "properties": {
                                    "gtw_id": {
                                        "type": "string"
                                    },
                                    "is_blocking": {
                                        "type": "integer"
                                    },
                                    "direction_id": {
                                        "type": "integer"
                                    },
                                    "mail_date": {
                                        "type": "date"
                                    },
                                    "subject": {
                                        "type": "string"
                                    },
                                    "email_from": {
                                        "type": "string"
                                    },
                                    "email_to": {
                                        "type": "string"
                                    },
                                    "email_cc": {
                                        "type": "string"
                                    },
                                    "email_bcc": {
                                        "type": "string"
                                    },
                                    "content": {
                                        "type": "string"
                                    },
                                    "content_html": {
                                        "type": "string"
                                    },
                                    "note": {
                                        "type": "string"
                                    },
                                    "mail_error_info": {
                                        "type": "string"
                                    },
                                    "outbound_category_id": {
                                        "type": "integer"
                                    },
                                    "outbound_category_detail_id": {
                                        "type": "integer"
                                    },
                                    "is_complete": {
                                        "type": "integer"
                                    },
                                    "uid": {
                                        "type": "string"
                                    },
                                    "user_id": {
                                        "type": "integer"
                                    },
                                    "campaign_id": {
                                        "type": "integer"
                                    },
                                    "customer_id": {
                                        "type": "integer"
                                    },
                                    "read_at": {
                                        "type": "date"
                                    },
                                    "read_by": {
                                        "type": "integer"
                                    },
                                    "reply_id": {
                                        "type": "integer"
                                    },
                                    "reply_at": {
                                        "type": "date"
                                    },
                                    "reply_by": {
                                        "type": "integer"
                                    },
                                    "forward_at": {
                                        "type": "date"
                                    },
                                    "forward_by": {
                                        "type": "integer"
                                    },
                                    "created_at": {
                                        "type": "date"
                                    },
                                    "created_by": {
                                        "type": "integer"
                                    },
                                    "updated_at": {
                                        "type": "date"
                                    },
                                    "updated_by": {
                                        "type": "integer"
                                    },
                                    "cust_name": {
                                        "type": "string"
                                    },
                                    "inbound_type_detail": {
                                        "type": "string"
                                    },
                                    "source_media_id": {
                                        "type": "integer"
                                    },
                                    "send_error_attempt": {
                                        "type": "string"
                                    },
                                    "is_spam": {
                                        "type": "integer"
                                    }
                                }
                            }
                        }
                    }
                },
                "description": "",
                "responses": {
                    "200": {
                        "description": "success"
                    },
                    "400": {
                        "description": "Bad Request"
                    },
                    "404": {
                        "description": "Not found."
                    }
                }
            }
        },
...
    },
    "schemes": [
        "http"
    ],
    "consumes": [
        "application/json"
    ],
    "produces": [
        "application/json"
    ]
}